MODLD-594: Resolve conflicts
